requirements.txt
pyyaml
pandas
jsonschema
matplotlib

